+++ src/test/java/org/reaktivity/specification/mqtt/ConnectionIT.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
-package org.reaktivity.specification.mqtt.streams;
+package org.reaktivity.specification.mqtt;
 
 import static java.util.concurrent.TimeUnit.SECONDS;
 import static org.junit.rules.RuleChain.outerRule;
